only 2 more tests to go - changed Specification from class to trait
