+++ core/shared/src/main/scala/org/qirx/littlespec/AbstractSpecification.scala
@@ -4,4 +4,4 @@ import scala.scalajs.js.annotation.{JSExportDescendentClasses, JSExportDescenden
 
 @JSExportDescendentObjects
 @JSExportDescendentClasses
-abstract trait Specification extends BaseSpecification
+abstract trait AbstractSpecification extends BaseSpecification
