+++ core/shared/src/test/scala/org/qirx/littlespec/sbt/SbtReporterSpec.scala
