+++ core/shared/src/test/scala/testUtils/NoJSExportSpecification.scala
@@ -2,4 +2,4 @@ package testUtils
 
 import org.qirx.littlespec.BaseSpecification
 
-abstract class NoJSExportSpecification extends BaseSpecification
+abstract trait NoJSExportSpecification extends BaseSpecification
